# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EmbedBase
  description: Local-first document embedding system with REST and MCP APIs
  version: 1.0.0
paths:
...
  /workspaces/{ws_id}/collections/{col_id}/documents:
    post:
      tags:
      - documents
      summary: Upload Document
      description: 'Upload and enqueue a document for ingestion.


        Accepted formats: PDF, TXT, and Markdown; DOCX and PPTX also require the docling

        parser backend (``parsers.pdf_backend: docling``). Set ``retention_days``
        (1-30) to

        auto-purge the document after that many days; omit it for a permanent document.'
      operationId: upload_document_workspaces__ws_id__collections__col_id__documents_post
      parameters:
      - name: ws_id
        in: path
        required: true
        schema:
          type: string
          title: Ws Id
      - name: col_id
        in: path
        required: true
        schema:
          type: string
          title: Col Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_workspaces__ws_id__collections__col_id__documents_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - documents
      summary: List Documents
      description: 'List documents in a collection: paginated, filtered, newest-first,
        with latest status.


        Returns ``{items, total, limit, offset}``. All filters are optional and AND-combined;

        ``filename`` is a case-insensitive substring, ``status`` is the latest ingestion
        status,

        ``indexed`` gates on stored chunks, and the ``*_size``/``*_after``/``*_before``
        bounds are

        inclusive. See :class:`DocumentListQuery` for the full set of query parameters.'
      operationId: list_documents_workspaces__ws_id__collections__col_id__documents_get
      parameters:
      - name: ws_id
        in: path
        required: true
        schema:
          type: string
          title: Ws Id
      - name: col_id
        in: path
        required: true
        schema:
          type: string
          title: Col Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: filename
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
      - name: file_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: indexed
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Indexed
      - name: embedding_model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Model
      - name: storage_backend
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Storage Backend
      - name: min_size
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Min Size
      - name: max_size
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Max Size
      - name: created_after
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Created After
      - name: created_before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Created Before
      - name: updated_after
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated After
      - name: updated_before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated Before
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Tag
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{ws_id}/collections/{col_id}/documents/{doc_id}/status:
    get:
      tags:
      - documents
      summary: Get Document Status
      description: Return the latest ingestion job status for a document.
      operationId: get_document_status_workspaces__ws_id__collections__col_id__documents__doc_id__status_get
      parameters:
      - name: ws_id
        in: path
        required: true
        schema:
          type: string
          title: Ws Id
      - name: col_id
        in: path
        required: true
        schema:
          type: string
          title: Col Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{ws_id}/collections/{col_id}/documents/{doc_id}:
    delete:
      tags:
      - documents
      summary: Delete Document
      description: Delete a document and enqueue async vector-store cleanup.
      operationId: delete_document_workspaces__ws_id__collections__col_id__documents__doc_id__delete
      parameters:
      - name: ws_id
        in: path
        required: true
        schema:
          type: string
          title: Ws Id
      - name: col_id
        in: path
        required: true
        schema:
          type: string
          title: Col Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /documents:
    post:
      tags:
      - documents
      summary: Upload Document Flat
      description: 'Upload a document by collection ID without the nested workspace
        path.


        Set ``retention_days`` (1-30) to auto-purge after that many days; omit for
        permanent.'
      operationId: upload_document_flat_documents_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_flat_documents_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /documents/{doc_id}/raw:
    get:
      tags:
      - documents
      summary: Get Document Raw
      description: 'Serve a document''s original bytes: inline FileResponse (local)
        or 302 to a

        presigned URL (S3), resolved from the document''s storage backend.'
      operationId: get_document_raw_documents__doc_id__raw_get
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /documents/{doc_id}:
    delete:
      tags:
      - documents
      summary: Delete Document Flat
      description: Delete a document by ID without the nested workspace/collection
        path.
      operationId: delete_document_flat_documents__doc_id__delete
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /documents/{doc_id}/reprocess:
    post:
      tags:
      - documents
      summary: Reprocess Document Flat
      description: 'Re-enqueue a document''s ingestion — the manual retry for a failed
        file. Reuses the stored

        bytes (nothing is re-uploaded) and surfaces a fresh pending job in the queue.'
      operationId: reprocess_document_flat_documents__doc_id__reprocess_post
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_flat_documents_post:
      properties:
        collection_id:
          type: string
          title: Collection Id
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        retention_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Retention Days
      type: object
      required:
      - collection_id
      - file
      title: Body_upload_document_flat_documents_post
    Body_upload_document_workspaces__ws_id__collections__col_id__documents_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        retention_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Retention Days
      type: object
      required:
      - file
      title: Body_upload_document_workspaces__ws_id__collections__col_id__documents_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...